+++ src/main/java/net/silentchaos512/wit/api/package-info.java
@@ -1,0 +1,9 @@
+@API(owner = "WIT", apiVersion = "2", provides = "WIT-API")
+@ParametersAreNonnullByDefault
+@MethodsReturnNonnullByDefault
+package net.silentchaos512.wit.api;
+
+import mcp.MethodsReturnNonnullByDefault;
+import net.minecraftforge.fml.common.API;
+
+import javax.annotation.ParametersAreNonnullByDefault;